feat(refs DPLAN-16766): Create separate directories for Vue component and twig
- to match the backend pattern that placed the controller in a separate procedurePhaseDefinition folder
- adjust the twig path in the controller
